src/storyboard/requirements.txt
flask 
pyjwt
pylint
jedi
flask_sqlalchemy
astroid
click
isort
itsdangerous
Jinja2
lazy-object-proxy
MarkupSafe
mccabe
SQLAlchemy 
marshmallow
psycopg2
parso
platformdirs
toml
werkzeug
wrapt
requests